requirements.txt
accelerate==0.22.0
add-parent-path==0.1.11
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
bitsandbytes==0.41.1
black==23.9.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
contourpy==1.1.1
cycler==0.11.0
diffusers==0.20.1
docker-pycreds==0.4.0
docstring-parser==0.15
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.12.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.33
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kanilog==0.3.3
kiwisolver==1.4.5
lit==16.0.6
logzero==1.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mc-bin-client==1.0.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.9
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
pathtools==0.1.2
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.17.0
pluggy==1.3.0
protobuf==4.24.2
psutil==5.9.5
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
pytorch-metric-learning==2.3.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
sentry-sdk==1.30.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stdlogging==0.16
streamlit==1.27.1
streamlit-state==0.1.0
sympy==1.12
tabulate==0.9.0
tap.py==3.1
tenacity==8.2.3
threadpoolctl==3.2.0
timm==0.9.6
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.7.1+cu110
torchvision==0.8.2+cu110
torchaudio==0.7.2
tornado==6.3.3
tqdm==4.66.1
transformers==4.32.1
triton==2.0.0
typed-argument-parser==1.8.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
validators==0.22.0
wandb==0.15.9
watchdog==3.0.0
yarl==1.9.2
zipp==3.16.2
